requirements.txt
pandas>=2.2
PyYAML>=6.0
requests>=2.31
xarray>=2024.2.0
zarr>=2.17
fsspec>=2024.2.0
scipy>=1.11
numpy>=1.26
